# openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "example": "invalid_request",
                "type": "string"
              },
              "field": {
                "example": "startsAt",
                "type": "string"
              },
              "message": {
                "example": "startsAt must be a future date and time",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Purchase": {
        "properties": {
          "amountMinor": {
            "example": 500000,
            "format": "int64",
            "type": "integer"
          },
          "checkoutUrl": {
            "example": "https://mock.paystack.local/checkout/018f2c1e-aaaa-7c3b-9d4e-2b8a1c5f7e90",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-09-23T17:58:00Z",
            "format": "date-time",
            "type": "string"
          },
          "eventId": {
            "example": "018f2c1e-9a11-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "id": {
            "example": "018f2c1e-aaaa-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "paidAt": {
            "example": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "initiated",
              "processing",
              "paid",
              "refunded",
              "failed"
            ],
            "example": "initiated",
            "type": "string"
          },
          "ticketId": {
            "example": "018f2c1e-cccc-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "userId": {
            "example": "018f2c1e-bbbb-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "userId",
          "amountMinor",
          "status",
          "ticketId",
          "createdAt"
        ],
        "type": "object"
      },
      "PurchaseTicketRequest": {
        "properties": {
          "idempotencyKey": {
            "type": "string"
          }
        },
        "required": [
          "idempotencyKey"
        ],
        "type": "object"
      },
...
      "Ticket": {
        "properties": {
          "eventId": {
            "example": "018f2c1e-9a11-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "id": {
            "example": "018f2c1e-cccc-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "issuedAt": {
            "example": "2026-09-23T18:01:00Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "reservationExpiresAt": {
            "example": "2026-09-23T18:08:00Z",
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "enum": [
              "temporarily_reserved",
              "reservation_expired",
              "issued",
              "revoked"
            ],
            "example": "issued",
            "type": "string"
          },
          "userId": {
            "example": "018f2c1e-bbbb-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "userId",
          "status",
          "reservationExpiresAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Opaque session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for paid live events: event lifecycle, ticket purchase, mocked payment, refunds, and payouts.",
    "title": "LIV DOT API",
    "version": "1.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/events/{id}/purchase": {
      "post": {
        "operationId": "purchaseTicket",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Reserve a ticket and start a payment",
        "tags": [
          "ticketing"
        ],
        "x-roles": [
          "user"
        ]
      }
    },
...
    "/api/tickets/{id}": {
      "get": {
        "operationId": "getTicket",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            },
            "description": "success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "not found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a ticket",
        "tags": [
          "ticketing"
        ],
        "x-roles": [
          "user"
        ]
      }
    },
    "/api/webhooks/payments": {
      "post": {
        "operationId": "receivePaymentWebhook",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "example": "received",
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "summary": "Receive a signed payment provider callback",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "local",
      "url": "http://localhost:8080"
    },
    {
      "description": "same origin",
      "url": "/"
    }
  ]
}